Created module validationDecorator.
This includes a directive which should be attached as an attribute on the form. The error handler service pushes the validation errors from service stack to a rootScope variable "validationErrors" if validation errors happen. For now the directive then attaches two data attrs with the valid state (true/false) and the error code returned by the API.