Finalized implementation of validation decorator
The validation errors from ServiceStack are now remapped into an object
based off the field it has errors for, and then an array of all the
validation errors for that field.

Each input is matched as before based off it's name and then a bootstrap
span error helper thingy is inserted with all the error texts that
ServiceStack returns.