The error handler is now a configurable provider
The error handler factory / http interceptor is now a provider so that
the inner error and success functions are configurable from the app.
This means that the error handler is no longer depending on rootScope as
the actual error handling is now delegated to the app using the error
handler.

Updated app.js to configure the error handler with the previously
implemented inner functions.